Fixed all the errors after compiling with the flags - option -d stopped working

+++ ls.h
@@ -54,7 +54,7 @@ bool isDisplayParent();
 void getBlocksAllocated();
 blkcnt_t calculateBlockSize(blkcnt_t blocks);
 char* generateHumanReadableSize(off_t);
-int numOfDigits(long long num);
+int numOfDigits(long num);
 void traverseDOption(struct f_non_dir non_dirs[], int n, struct f_dir n_dirs[], int m);
 char* checkNonPrintChars(char*);
 void getPrintDetails(FTSENT *ftsent, struct perttyPrint *pPrint);
